+++ src/concurrency/mod.rs
@@ -1,6 +1,8 @@
 pub mod data_race;
 mod range_object_map;
+#[macro_use]
 pub mod sync;
+pub mod init_once;
 pub mod thread;
 mod vector_clock;
 pub mod weak_memory;
